+++ workspace/solutions/ecmascript/src/s1/s4/s9/s9/s_1499_test.ts
@@ -1,0 +1,8 @@
+import test from "bun:test";
+
+import { t } from "@/utils/test";
+
+await t(
+    test,
+    import.meta.dir,
+);