+++ workspace/solutions/kotlin/lib/src/test/kotlin/s1/s4/s9/s9/s_1499_test.kt
@@ -1,0 +1,16 @@
+package s1.s4.s9.s9
+
+import utils.T
+import kotlin.test.Test
+import s1.s4.s9.s9.s_1499_00.Solution as s_1499_00
+
+class s_1499_test {
+    @Test
+    fun test() {
+        val t = T(
+            this::class,
+            s_1499_00::class,
+        )
+        t.run()
+    }
+}