+++ workspace/solutions/rust/src/s1/mod.rs
@@ -1,1 +1,2 @@
+pub mod s4;
 pub mod s8;