+++ workspace/solutions/rust/src/s1/s4/s9/s9/mod.rs
@@ -1,0 +1,5 @@
+pub mod s_1499_00;
+pub mod s_1499_01;
+pub mod s_1499_test;
+
+pub type TSolution = fn(Vec<Vec<i32>>, i32) -> i32;